ZA: Move the views related with location in views.py to their own file
The imports that were only used by these views were moved as well,
and the CONSTITUENCY_OFFICE_PLACE_KIND_SLUGS constant
and WardCouncillorAPIDown expection, as they are only used by
the classes included in this file.
