+++ pombola/south_africa/views/__init__.py
@@ -1,3 +1,4 @@
 from .constants import *
+from .geolocalization import *
 from .home import *
 from .views import *
